Move Unify.{h,cc} to opencog/unify directory
