+++ opencog/rule-engine/backwardchainer/BackwardChainer.cc
@@ -31,7 +31,7 @@
 
 #include <opencog/atomutils/FindUtils.h>
 #include <opencog/atomutils/Substitutor.h>
-#include <opencog/atomutils/Unify.h>
+#include <opencog/unify/Unify.h>
 #include <opencog/atomutils/TypeUtils.h>
 #include <opencog/atoms/pattern/PatternLink.h>
 #include <opencog/atoms/pattern/BindLink.h>
