+++ tests/CMakeLists.txt
@@ -21,6 +21,7 @@ IF (CXXTEST_FOUND)
 	ADD_SUBDIRECTORY (atomspace)
 	ADD_SUBDIRECTORY (atoms)
 	ADD_SUBDIRECTORY (atomutils)
+	ADD_SUBDIRECTORY (unify)
 	ADD_SUBDIRECTORY (atomspaceutils)
 	ADD_SUBDIRECTORY (attentionbank)
