+++ tests/atomutils/CMakeLists.txt
@@ -11,4 +11,3 @@ ADD_CXXTEST(AtomUtilsUTest)
 ADD_CXXTEST(FindUtilsUTest)
 ADD_CXXTEST(FuzzyUTest)
 ADD_CXXTEST(NeighborUTest)
-ADD_CXXTEST(UnifyUTest)
