+++ tests/unify/CMakeLists.txt
@@ -1,0 +1,10 @@
+LINK_LIBRARIES(
+	unify
+	atomspace
+	clearbox
+)
+IF (HAVE_GUILE)
+	LINK_LIBRARIES(smob)
+ENDIF (HAVE_GUILE)
+
+ADD_CXXTEST(UnifyUTest)
